Struct ink_lang_codegen::generator::cross_calling::CrossCalling

pub struct CrossCalling<'a> {
    contract: &'a Contract,
}

Generates code for using this ink! contract as a dependency.

Fields
contract: &'a Contract
Implementations

impl<'_> CrossCalling<'_>

pub(in generator::cross_calling) fn generate_cfg(
&self
) -> Option<TokenStream2>

Generates code for conditionally compiling code only if the contract is compiled as dependency.

pub(in generator::cross_calling) fn generate_storage(
&self
) -> TokenStream2

Generates code for the ink! storage struct for cross-calling purposes.

Note

This always consists of a single AccountId and can be viewed as a reference to a live smart contract instance of the same type. It will forward all calls via ink!'s provided cross-calling infrastructure automatically over the chain.

pub(in generator::cross_calling) fn generate_standard_impls(
&self
) -> TokenStream2

Generates code for the trait implementations required to make the generated ink! storage struct for cross-calling work out of the box for the cross-calling infrastructure.

pub(in generator::cross_calling) fn generate_arg_list<'a, Args>(
args: Args
) -> TokenStream2 where
Args: IntoIterator<Item = &'a Type>,
<Args as IntoIterator>::IntoIter: DoubleEndedIterator,

Builds up the [ink_env::call::ArgumentList] type structure for the given types.

pub(in generator::cross_calling) fn call_forwarder_ident(
) -> Ident

Returns the identifier for the generated call forwarder utility.

pub(in generator::cross_calling) fn generate_call_forwarder_trait_ghost_message(
message: CallableWithSelector<'_, Message>
) -> TokenStream2

pub(in generator::cross_calling) fn generate_call_forwarder_trait_proper_message(
message: CallableWithSelector<'_, Message>
) -> TokenStream2

pub(in generator::cross_calling) fn generate_call_forwarder_trait_message(
mutable: bool,
message: CallableWithSelector<'_, Message>
) -> TokenStream2

Generates code for a single call forwarder trait message.

The mutable parameter indicates whether only read-only (false) or write-only (true) messages shall be valid calls. For non valid messages an invalid implementation is provided so that actually calling those will result in a compiler or linker error.

pub(in generator::cross_calling) fn generate_call_forwarder_trait_constructor(
constructor: CallableWithSelector<'_, Constructor>
) -> TokenStream2

Generates code for a single call forwarder trait constructor.

Note that constructors never need to be forwarded and that we only provide their implementations to satisfy the implementation block. We generally try to generate code in a way that actually calling those constructors will result in a compiler or linker error.

pub(in generator::cross_calling) fn generate_call_forwarder_trait_impl_block(
&self,
mutable: bool,
item_impl: &ItemImpl
) -> TokenStream2

Generates code for a single call forwarder trait implementation block.

The mutable parameter indicates whether only read-only (false) or write-only (true) messages and constructors are to be considered.

pub(in generator::cross_calling) fn generate_call_forwarder_inherent_message(
message: CallableWithSelector<'_, Message>
) -> TokenStream2

pub(in generator::cross_calling) fn generate_call_forwarder_inherent_impl_block(
&self,
mutable: bool,
item_impl: &ItemImpl
) -> TokenStream2

Generates code for a single call forwarder inherent implementation block.

The mutable parameter indicates whether only read-only (false) or write-only (true) messages and constructors are to be considered.

pub(in generator::cross_calling) fn generate_call_forwarder_impl_blocks(
&self,
mutable: bool
) -> TokenStream2

Generates code for the call forwarder implementation blocks.

The mutable parameter indicates whether only read-only (false) or write-only (true) messages and constructors are to be considered.

pub(in generator::cross_calling) fn generate_call_forwarders(
&self
) -> TokenStream2

Generates code for the call forwarder utility struct.

pub(in generator::cross_calling) fn generate_trait_impl_block_message(
&self,
message: CallableWithSelector<'_, Message>
) -> TokenStream2

Generates the code to allow short-hand cross-chain contract calls for messages.

pub(in generator::cross_calling) fn generate_trait_impl_block_constructor(
constructor: CallableWithSelector<'_, Constructor>
) -> TokenStream2

Generates the code to allow cross-chain contract calls for trait constructors.

pub(in generator::cross_calling) fn generate_trait_impl_block(
&self,
impl_block: &ItemImpl
) -> TokenStream2

pub(in generator::cross_calling) fn generate_inherent_impl_block_constructor(
constructor: CallableWithSelector<'_, Constructor>
) -> TokenStream2

Generates the code to allow short-hand cross-chain contract calls for constructors.

Note

For constructors this is the only way they are able to be called.

pub(in generator::cross_calling) fn generate_inherent_impl_block_message(
&self,
message: CallableWithSelector<'_, Message>
) -> TokenStream2

Generates the code to allow short-hand cross-chain contract calls for messages.

pub(in generator::cross_calling) fn generate_inherent_impl_block(
&self,
impl_block: &ItemImpl
) -> TokenStream2

pub(in generator::cross_calling) fn generate_impl_blocks(
&self
) -> TokenStream2