Struct ink_lang_codegen::generator::dispatch::Dispatch

pub struct Dispatch<'a> {
    contract: &'a Contract,
}

Generates code for the message and constructor dispatcher.

This code efficiently selects the dispatched ink! constructor or message by inspecting the first four bytes (selector) of the given input bytes.

As this happens on every contract execution this code must be highly optimized. For that purpose a so-called dispatch enum is being generated that has a specialized scale::Decode implementation taking the first four bytes of the input stream in order to identify the enum variant that it is going to produce out of the rest of the input buffer.

The rest of the input buffer is then automatically decoded directly into the expected input types of the respective ink! constructor or message.

Fields
contract: &'a Contract
Implementations

impl<'_> Dispatch<'_>

pub(in generator::dispatch) fn generate_entry_points(
&self
) -> TokenStream2

Generates the static ink! contract entry points.

Note

Those are expected to exist by the smart contracts host module. They guide the dispatch, set-up and tear-down of a smart contract.

pub(in generator::dispatch) fn generate_dispatch_using_mode(
&self
) -> TokenStream2

Generates the DispatchUsingMode trait implementation to guide contract dispatch.

pub(in generator::dispatch) fn dispatch_trait_impl_namespace(
kind: CallableKind
) -> Ident

Returns the generated ink! namespace identifier for the given callable kind.

pub(in generator::dispatch) fn generate_trait_impl_namespaces(
&self
) -> TokenStream2

Generates utility types to emulate namespaces to disambiguate dispatch trait implementations for ink! messages and ink! constructors with overlapping selectors.

pub(in generator::dispatch) fn generate_trait_impls_for_callable<C>(
&self,
cws: CallableWithSelector<'_, C>
) -> TokenStream2 where
C: Callable + ToTokens,

Generates code for the dispatch trait impls for a generic ink! callable.

pub(in generator::dispatch) fn generate_input_bindings<C>(
callable: &C
) -> (Vec<Ident>, TokenStream2) where
C: Callable,

Returns a tuple of:

Vector over the generated identifier bindings (__ink_binding_N) for all inputs.

TokenStream representing the binding identifiers as tuple (for >=2 inputs), as single identifier (for exactly one input) or as wildcard (_) if there are no input bindings.

Examples

No inputs:

( vec![],
  quote! { _ } )

Exactly one input:

( vec![__ink_binding_0],
  quote! { __ink_binding_0 } )

Multiple (>=2) inputs:

( vec![__ink_binding_0, __ink_binding_1, /* ... */],
  quote! { (__ink_binding_0, __ink_binding_1, ..) } )

pub(in generator::dispatch) fn generate_trait_impls_for_message(
&self,
cws: CallableWithSelector<'_, Message>
) -> TokenStream2

Generates all the dispatch trait implementations for the given ink! message.

pub(in generator::dispatch) fn generate_trait_impls_for_constructor(
&self,
cws: CallableWithSelector<'_, Constructor>
) -> TokenStream2

Generates all the dispatch trait implementations for the given ink! constructor.

pub(in generator::dispatch) fn generate_dispatch_trait_impls(
&self
) -> TokenStream2

Generate all dispatch trait implementations for ink! messages and ink! constructors.

pub(in generator::dispatch) fn generate_dispatch_variant_ident<C>(
&self,
cws: CallableWithSelector<'_, C>
) -> Ident where
C: Callable,

Generates variant identifiers for the generated dispatch enum.

Since we want to avoid generating random names we generate identifiers in terms of the selectors of the associated ink! messages or constructors.

Example

Given prefix of "Message" and selector with bytes 0xDEADBEEF we generate the following idenfitier: __ink_Message_0xDEADBEEF

This way it is clear that this is an ink! generated identifier and even encodes the unique selector bytes to make the identifier unique.

pub(in generator::dispatch) fn generate_dispatch_variant_decode<C>(
&self,
cws: CallableWithSelector<'_, C>
) -> TokenStream2 where
C: Callable,

Generates one match arm of the dispatch scale::Decode implementation.

Note

There is one match arm per ink! constructor or message for the dispatch scale::Decode implementation.

pub(in generator::dispatch) fn generate_dispatch_variant_arm<C>(
&self,
cws: CallableWithSelector<'_, C>
) -> TokenStream2 where
C: Callable,

Generates one match arm of the dispatch variant enum.

Note

There is one match arm per ink! constructor or message for the dispatch scale::Decode implementation.

pub(in generator::dispatch) fn all_messages_deny_payment(
&self
) -> bool

Returns true if all ink! messages of self deny payments.

Note

This information is used to produce better code in this scenario.

pub(in generator::dispatch) fn generate_dispatch_execute_message_arm(
&self,
cws: CallableWithSelector<'_, Message>
) -> TokenStream2

Generates one match arm of the dispatch message for the execute implementation.

Note

This is basically the code per ink! message that is going to be executed after the dispatch has already taken place.

pub(in generator::dispatch) fn contract_messages(
&self
) -> impl Iterator<Item = CallableWithSelector<'_, Message>>

Returns an iterator over all ink! messages of the ink! contract.

pub(in generator::dispatch) fn generate_message_dispatch_enum(
&self
) -> TokenStream2

Generates the entire dispatch variant enum for all ink! messages.

pub(in generator::dispatch) fn generate_dispatch_execute_constructor_arm(
&self,
cws: CallableWithSelector<'_, Constructor>
) -> TokenStream2

Generates one match arm of the dispatch constructor for the execute implementation.

Note

This is basically the code per ink! constructor that is going to be executed after the dispatch has already taken place.

pub(in generator::dispatch) fn contract_constructors(
&self
) -> impl Iterator<Item = CallableWithSelector<'_, Constructor>>

Returns an iterator over all ink! constructors of the ink! contract.

pub(in generator::dispatch) fn generate_constructor_dispatch_enum(
&self
) -> TokenStream2

Generates the entire dispatch variant enum for all ink! messages.